Updated Facility table
Added the following columns to the facility table:

-email
-phone
-city
-province
-country
-postal code

Rememeber run the masterscript to get the db changes. Also remember that you may to update the model in the project. To do this, open the Model1.edmx file. Then right click anywhere on that page and then select the option "Update model from database"